Add lib support for new architecture
