+++ ios/MlkitOcr.h
@@ -1,5 +1,14 @@
 #import <React/RCTBridgeModule.h>
 
-@interface MlkitOcr : NSObject <RCTBridgeModule>
+#ifdef RCT_NEW_ARCH_ENABLED
 
+#import "MlkitOcrSpec.h"
+@interface MlkitOcr : NSObject <MlkitOcrSpec>
 @end
+
+#else
+
+@interface MlkitOcr : NSObject <RCTBridgeModule>
+@end
+
+#endif
